Move HLSLTypes to Utility
