+++ test/Private/D3D12/Shader/ShaderBindingTests.cpp
@@ -1,7 +1,7 @@
 
 #include <Platform.h>
 #include <D3D12/Shader/ShaderBinding.h>
-#include <Framework/HLSLTypes.h>
+#include <Utility/HLSLTypes.h>
 
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/catch_template_test_macros.hpp>
