+++ test/Private/Utility/HLSLTypesTests.cpp
@@ -1,4 +1,4 @@
-#include "Framework/HLSLTypes.h"
+#include <Utility/HLSLTypes.h>
 
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/catch_template_test_macros.hpp>
